requirements.txt
pyperclip
requests
python-dotenv
mss
openai
Pillow
